requirement.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
langchain
langchain-groq